MaskCreator, Metronome: fix compiler warnings

+++ src/modules/Metronome/Metronome.cpp
@@ -6,7 +6,7 @@ using namespace std;
 Metronome::Metronome(Configuration config, std::vector<Detector*> detectors)
     : Module(std::move(config), std::move(detectors)) {
 
-    m_eventLength = m_config.get<double>("eventLength", Units::convert(10, "us"));
+    m_eventLength = m_config.get<double>("eventLength", static_cast<double>(Units::convert(10, "us")));
 }
 
 void Metronome::initialise() {
